### 2023-12-01 #6103 @editorialbot
[PRE REVIEW]: InvertibleNetworks.jl: A Julia package for scalable normalizing flows
**Submitting author:** @rafaelorozco (<a href="http://orcid.org/0000-0003-0917-2442">Rafael Orozco</a>)
**Repository:** https://github.com/slimgroup/InvertibleNetworks.jl
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** v2.2.6
