### 2023-12-01 #6103 @rafaelorozco
Hello @danielskatz thank you for the triage help!

I will work on the DOI warnings. I would like to suggest Nando Hegemann as a reviewer. Their github handle seems to be Nando-Hegemann or Nando-Farchmin

 Thank you!
