### 2024-03-23 8bac1b8
tests working cpu julia 10.0
