### 2024-03-23 9deb875
add julia 10 to workflow
