+++ .github/workflows/runtests.yml
@@ -19,7 +19,7 @@ jobs:
       fail-fast: false
 
       matrix:
-        version: ['1.6', '1.7', '1.8']
+        version: ['1.6', '1.7', '1.8', '1.10']
         tests: ["basics", "layers", "networks"]
         os: [ubuntu-latest]
 
