### 2024-03-24 035ca66
mean grad test over ratios
