### 2024-03-24 da34439
mean grad test for hint cond
