### 2024-03-24 ec34cc7
add gpu tests for cond glow
