### 2024-03-26 bb4d611
self runner only for single gpu test
