+++ test/test_networks/test_conditional_glow_network.jl
@@ -4,6 +4,7 @@
 
 using InvertibleNetworks, LinearAlgebra, Test,Flux, Random
 device = InvertibleNetworks.CUDA.functional() ? gpu : cpu
+(device == gpu) && println("Testing on GPU"); 
 
 # Random seed
 Random.seed!(3);
