+++ test/test_networks/test_glow.jl
@@ -2,7 +2,16 @@
 # Author: Philipp Witte, pwitte3@gatech.edu
 # Date: January 2020
 
-using InvertibleNetworks, LinearAlgebra, Test, Random
+using InvertibleNetworks, LinearAlgebra, Test, Flux, Random
+device = InvertibleNetworks.CUDA.functional() ? gpu : cpu
+(device == gpu) && println("Testing on GPU"); 
+
+mean(x) = sum(x)/length(x)
+
+hfactor=8f-1
+expected_f1 = 1f0 / hfactor
+expected_f2 = 1f0 / hfactor ^2f0
+
 
 # Random seed
 m = MersenneTwister()
@@ -23,8 +32,8 @@ for logdet_bool = [true,false] #logdet is not tested in Jacobian yet.
             println("Testing Glow with dimensions=$(N) logdet=$(logdet_bool) and split_scales=$(split_scales)")
             
             # Network and input
-            G = NetworkGlow(n_in, n_hidden, L, K;logdet=logdet_bool, split_scales=split_scales, ndims=length(N))
-            X = rand(m,Float32, N..., n_in, batchsize)
+            G = NetworkGlow(n_in, n_hidden, L, K;logdet=logdet_bool, split_scales=split_scales, ndims=length(N))|> device
+            X = rand(m,Float32, N..., n_in, batchsize)|> device
 
             # Invertibility
             if logdet_bool
@@ -75,9 +84,9 @@ for logdet_bool = [true,false] #logdet is not tested in Jacobian yet.
             end
 
             # Gradient test w.r.t. input
-            G = NetworkGlow(n_in, n_hidden, L, K;logdet=logdet_bool, split_scales=split_scales, ndims=length(N))
-            X = rand(m,Float32, N..., n_in, batchsize)
-            X0 = rand(m,Float32, N..., n_in, batchsize)
+            G = NetworkGlow(n_in, n_hidden, L, K;logdet=logdet_bool, split_scales=split_scales, ndims=length(N))|> device
+            X = rand(m,Float32, N..., n_in, batchsize)|> device
+            X0 = rand(m,Float32, N..., n_in, batchsize)|> device
             dX = X - X0
 
             f0, ΔX = loss(G, X0)[1:2]
@@ -98,36 +107,48 @@ for logdet_bool = [true,false] #logdet is not tested in Jacobian yet.
             @test isapprox(err1[end] / (err1[1]/2^(maxiter-1)), 1f0; atol=1f1)
             @test isapprox(err2[end] / (err2[1]/4^(maxiter-1)), 1f0; atol=1f1)
 
-            # Gradient test w.r.t. parameters
-            X = rand(m,Float32, N..., n_in, batchsize)
-            G = NetworkGlow(n_in, n_hidden, L, K; logdet=logdet_bool, split_scales=split_scales, ndims=length(N))
-            G0 = NetworkGlow(n_in, n_hidden, L, K;logdet=logdet_bool, split_scales=split_scales, ndims=length(N))
-            Gini = deepcopy(G0)
+            num_attempts = 3
+            results_1 = []
+            results_2 = []
+            for i in 1:num_attempts
+                Random.seed!(i)
 
-            # Test one parameter from residual block and 1x1 conv
-            dW = G.CL[1,1].RB.W1.data - G0.CL[1,1].RB.W1.data
-            dv = G.CL[1,1].C.v1.data - G0.CL[1,1].C.v1.data
+                # Gradient test w.r.t. parameters
+                X = rand(m,Float32, N..., n_in, batchsize)|> device
+                G = NetworkGlow(n_in, n_hidden, L, K; logdet=logdet_bool, split_scales=split_scales, ndims=length(N))|> device
[…]
