### 2024-03-26 bbd4519
clean name self hosted
