### 2024-03-30 #6554 @editorialbot
[REVIEW]: InvertibleNetworks.jl: A Julia package for scalable normalizing flows
**Submitting author:** @rafaelorozco (<a href="http://orcid.org/0000-0003-0917-2442">Rafael Orozco</a>)
**Repository:** https://github.com/slimgroup/InvertibleNetworks.jl
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** 2.2.9
