### 2024-03-30 #6554 @drvinceknight
👋🏼 @rafaelorozco @aurorarossi & @Nando-Hegemann this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

Please feel free to ping me (@drvinceknight ) if you have any questions/concerns.
