+++ docs/paper/paper.md
@@ -40,37 +40,37 @@ bibliography: paper.bib
 
 # Summary
 
-Normalizing flows is a density estimation method that provides efficient exact likelihood estimation and sampling [@dinh2014nice] from high dimensional distributions. This method depends on the use of the change of variables formula which requires an invertible transform. Thus normalizing flow architectures are built to be invertible by design [@dinh2014nice]. In theory, the invertibility of architectures constrains the expressiveness but the use of coupling layers allows normalizing flows to exploit the power of arbitrary neural networks that need not be invertible [@dinh2016density] and layer invertibility means if properly implemented many layers can be stacked to increase expressiveness without creating a training memory bottleneck.  
+Normalizing flows is a density estimation method that provides efficient exact likelihood estimation and sampling [@dinh2014nice] from high-dimensional distributions. This method depends on the use of the change of variables formula, which requires an invertible transform. Thus normalizing flow architectures are built to be invertible by design [@dinh2014nice]. In theory, the invertibility of architectures constrains the expressiveness, but the use of coupling layers allows normalizing flows to exploit the power of arbitrary neural networks, which do not need to be invertible, [@dinh2016density] and layer invertibility means that, if properly implemented, many layers can be stacked to increase expressiveness without creating a training memory bottleneck.  
 
-The package we present, InvertibleNetworks.jl, is a pure Julia [@bezanson2012julia] implementation of normalizing flows. We have implemented many relevant neural network layers, including GLOW 1x1 invertible convolutions [@kingma2018glow], affine/additive coupling layers [@dinh2014nice], Haar wavelet multiscale transforms [@haar1909theorie] and Hierarchical invertible neural transport (HINT) [@kruse2021hint] among others. These modular layers are easily composed and modified to create different types of normalizing flows. As starting points, we have implemented RealNVP, GLOW, HINT, Hyperbolic networks [@lensink2022fully] and their conditional counterparts for users to quickly implement their individual applications. 
+The package we present, InvertibleNetworks.jl, is a pure Julia [@bezanson2012julia] implementation of normalizing flows. We have implemented many relevant neural network layers, including GLOW 1x1 invertible convolutions [@kingma2018glow], affine/additive coupling layers [@dinh2014nice], Haar wavelet multiscale transforms [@haar1909theorie], and Hierarchical invertible neural transport (HINT) [@kruse2021hint], among others. These modular layers can be easily composed and modified to create different types of normalizing flows. As starting points, we have implemented RealNVP, GLOW, HINT, Hyperbolic networks [@lensink2022fully] and their conditional counterparts for users to quickly implement their individual applications. 
 
 # Statement of need
 
-This software package focuses on memory efficiency. The promise of neural networks is in learning high-dimensional distributions from examples thus normalizing flow packages should allow easy application to large dimensional inputs such as images or 3D volumes. Interestingly, the invertibility of normalizing flows naturally alleviates memory concerns since intermediate networks activations can be recomputed instead of saved in memory, greatly reducing the memory needed during backpropagation. The problem is that directly implementing normalizing flows in automatic differentiation frameworks such as PyTorch [@paszke2017automatic] will not automatically exploit this invertibility. The available packages for normalizing flows such as: nflows [@nflows], normflows [@stimper2023normflows] and FrEIA [@freia] are built depending on automatic differentiation frameworks thus do not exploit invertibility for memory efficiently. 
[…]
