### 2024-04-02 slimgroup/InvertibleNetworks.jl#106 @aurorarossi
Improve visibility of contribution guidelines
Ref openjournals/joss-reviews#6554
A requirement of JOSS is to have clear guidelines for third parties who wish to:
  - Contribute to the software
 - Report issues or problems with the software

  - Seek support

Namely, the `CONTRIBUTING.md` file is missing.
Here are some examples:
- https://github.com/JuliaGraphs/Graphs.jl/blob/master/CONTRIBUTING.md
- https://github.com/FluxML/Flux.jl/blob/master/CONTRIBUTING.md
- https://github.com/JuliaGraphs/GraphsOptim.jl/blob/main/CONTRIBUTING.md
