### 2024-04-02 907c391
New Julia ref in paper.bib
