+++ docs/paper/paper.bib
@@ -7,11 +7,16 @@
   year={2022}
 }
 
-@article{bezanson2012julia,
-  title={Julia: A fast dynamic language for technical computing},
-  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
-  journal={arXiv preprint arXiv:1209.5145},
-  year={2012}
+@article{bezanson2017julia,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM},
+  url={https://doi.org/10.1137/141000671}
 }
 
 @article{peters2019symmetric,
