### 2024-04-02 c478401
Update Julia ref paper.md
