+++ docs/paper/paper.bib
@@ -1,5 +1,5 @@
 @article{peters2022point,
-  title={Point-TO-SET DISTANCE FUNCTIONS FOR OUTPUT-CONSTRAINED NEURAL NETWORKS.},
+  title={Point-to-set distance functions for output-constrained neural networks.},
   author={Peters, Bas},
   journal={Journal of Applied \& Numerical Optimization},
   volume={4},
