### 2024-04-02 slimgroup/InvertibleNetworks.jl#106 @mloubout
We do have a

- contributing
- code of conduct

These we have set up organization-wide. You can see it in the bottom corner when opening an issue and PR where it's supposed to be.
