### 2024-04-02 slimgroup/InvertibleNetworks.jl#106 @aurorarossi
Ah ok, I did not see it. I saw the code of conduct, which is visible near the README, but the contributing file is not as visible. It is ok that it is the same as the organization one, but just looking at the repository, if someone wants to open a PR and has not opened an issue, they will miss that information. I think you could create a section in the README that points this out.
