### 2024-04-18 #6554 @aurorarossi
@rafaelorozco There are many examples in your repository, so it would be nice if you could also add the code to reproduce the benchmark plots of the paper using Pytorch.
