### 2024-05-02 a11600c
change vers and add cudnn dep
