+++ Project.toml
@@ -13,13 +13,14 @@ NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Wavelets = "29a6e085-ba6d-5f35-a997-948ac2efa89a"
+cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
 
 [compat]
-CUDA = "1, 2, 3, 4"
+CUDA = "1, 2, 3, 4, 5"
 ChainRulesCore = "0.8, 0.9, 0.10, 1"
-Flux = "0.11, 0.12, 0.13"
+Flux = "0.11, 0.12, 0.13, 0.14"
 JOLI = "0.7, 0.8"
-NNlib = "0.7, 0.8"
+NNlib = "0.7, 0.8, 0.9"
 TimerOutputs = "0.5"
 Wavelets = "0.9, 0.10"
 julia = "1"
