### 2024-05-10 1e5016f
add conditional mnist example
