+++ README.md
@@ -16,7 +16,6 @@ Building blocks for invertible neural networks in the [Julia] programming langua
 
 ## Installation
 
-
 InvertibleNetworks is registered and can be added like any standard Julia package with the command:
 
 ```
@@ -24,37 +23,13 @@ InvertibleNetworks is registered and can be added like any standard Julia packag
 ```
 
 
-## Papers
+## Uncertainty-aware image reconstruction
 
-The following publications use [InvertibleNetworks.jl]:
 
-- **["Reliable amortized variational inference with physics-based latent distribution correction"]**
-    - paper: [https://arxiv.org/abs/2207.11640](https://arxiv.org/abs/2207.11640)
-    - [presentation](https://slim.gatech.edu/Publications/Public/Submitted/2022/siahkoohi2022ravi/slides.pdf)
-    - code: [ReliableAVI.jl]
+Due to its memory scaling InvertibleNetworks.jl has been particularily successful at Bayesian posterior sampling with simulation-based inference. To get started with this application please refer to a simple example ([Conditional sampling for MNSIT inpainting](https://github.com/slimgroup/InvertibleNetworks.jl/tree/master/examples/applications/application_conditional_mnist_inpainting.jl)) but please use this on your application and please reach out to us if you run into any trouble. 
 
-- **["Learning by example: fast reliability-aware seismic imaging with normalizing flows"]**
-    - paper: [https://arxiv.org/abs/2104.06255](https://arxiv.org/abs/2104.06255)
-    - [presentation](https://slim.gatech.edu/Publications/Public/Conferences/KAUST/2021/siahkoohi2021EarthMLfar/siahkoohi2021EarthMLfar.pdf)
-    - code: [ReliabilityAwareImaging.jl]
+![mnist_sampling_cond](docs/src/figures/mnist_sampling_cond.png)
 
-- **["Enabling uncertainty quantification for seismic data pre-processing using normalizing flows (NF)—an interpolation example"]**
-    - [paper](https://slim.gatech.edu/Publications/Public/Conferences/SEG/2021/kumar2021SEGeuq/kumar2021SEGeuq.pdf)
-    - code: [WavefieldRecoveryUQ.jl]
-
-- **["Preconditioned training of normalizing flows for variational inference in inverse problems"]**
-    - paper: [https://arxiv.org/abs/2101.03709](https://arxiv.org/abs/2101.03709)
-    - [presentation](https://slim.gatech.edu/Publications/Public/Conferences/AABI/2021/siahkoohi2021AABIpto/siahkoohi2021AABIpto_pres.pdf)
-    - code: [FastApproximateInference.jl]
-
-- **["Parameterizing uncertainty by deep invertible networks, an application to reservoir characterization"]**
-    - paper: [https://arxiv.org/abs/2004.07871](https://arxiv.org/abs/2004.07871)
-    - [presentation](https://slim.gatech.edu/Publications/Public/Conferences/SEG/2020/rizzuti2020SEGuqavp/rizzuti2020SEGuqavp_pres.pdf)
-    - code: [https://github.com/slimgroup/Software.SEG2020](https://github.com/slimgroup/Software.SEG2020)
-
-- **["Generalized Minkowski sets for the regularization of inverse problems"]**
-    - paper: [http://arxiv.org/abs/1903.03942](http://arxiv.org/abs/1903.03942)
-    - code: [SetIntersectionProjection.jl]
 
 ## Building blocks
 
@@ -112,26 +87,56 @@ AN = ActNorm(k; logdet=true) |> gpu
[…]
