+++ Project.toml
@@ -17,6 +17,7 @@ Wavelets = "29a6e085-ba6d-5f35-a997-948ac2efa89a"
 
 [compat]
 CUDA = "1, 2, 3, 4, 5"
+cuDNN = "1"
 ChainRulesCore = "0.8, 0.9, 0.10, 1"
 Flux = "0.11, 0.12, 0.13, 0.14"
 JOLI = "0.7, 0.8"
