### 2024-05-22 #6554 @Nando-Hegemann
Here are some of the remarks I have:

**General Checks -> Reproducibility**:
I'm not able to reproduce the results of the paper (or run any of the other examples). I'm aware that this is due to my lack of familiarity with the Julia language, but maybe you could include a simple section on "How to run the MNIST example" (simple command line instructions) in the README/Doc as well as a description on how to set up an environment to execute the examples/tests. Right now, after freshly installing Julia and running the scripts/tests I simply get LoadError(s).

**Functionality -> Functionality** & **Functionality -> Performance**:
I agree with @aurorarossi that a script reproducing the plots of the paper would be good. A simple (python) script and respective environment for the python setup used to produce the results would be a good addition.

**Documentation -> Example usage**:
1. You have a lot of example scripts of which only a few are included in the online documentation (which is fine). However, the examples in the online documentation on provide the source code of the scripts and an image. I think providing more context to the examples (what is the setting, why is it interesting as an example) and describing how the individual steps work and what the output picture depicts would be good.
2. Currently you copy-pasted source code from your scripts to the examples in the doc. To decrease maintenance issues, I'd suggest referencing (certain lines of) the Julia scripts in the doc, so that the doc always shows source code from executable files. This way your doc and the example scripts don't run out of sync if you change something.

With .rst files you could do something like
```
.. literalinclude:: ../examples/applications/file_name.jl
  :language: julia
  :lines: 8-9
```

**Documentation -> Community Guidelines**:
Again, I agree with @aurorarossi to include a clearly visible link to your general contribution guidelines in your README.

**Software Paper -> State of the field**:
In the paper you describe that your software performs faster then other packages such as FrEIA or normflows, but I'd suggest adding some additional notes on what separates this software from other packages. For example you could focus on the following:
1. Is there a reason to implement Invertible Networks in Julia rather than in python?
2. Does your software include algorithms/architectures/training routines not implemented anywhere else?
3. Does your software combine established algorithms from multiple other packages, thus providing a more versatile and easy to use framework?
In other words, you should reason why you implemented a new software package. If you mainly compute gradients analytically instead of using autograd one could argue that it would be better to integrate these computational improvements in other already existing packages. Also, are there any other Julia packages for INNs? If so they should be mentioned and differences should be explained as well.

**Software Paper -> Quality of writing**:
I have only some minor formulation suggestions, which I'm going to add to https://github.com/slimgroup/InvertibleNetworks.jl/pull/105 soon.
