+++ src/networks/summarized_net.jl
@@ -34,23 +34,30 @@ export SummarizedNet
 struct SummarizedNet <: InvertibleNetwork
 	cond_net::InvertibleNetwork
 	sum_net
+    squeezer::Squeezer
 end
 
 @Flux.functor SummarizedNet
 
 # Forward pass 
 function forward(X::AbstractArray{T, N}, Y::AbstractArray{T, N}, S::SummarizedNet) where {T, N}
+    X = S.squeezer.forward(X)
+    Y = S.squeezer.forward(Y)
     S.cond_net(X,  S.sum_net(Y))
+    #X, Y, lgdet = S.cond_net(X,  S.sum_net(Y))
+    #S.squeezer.inverse(X), S.squeezer.inverse(Y), lgdet
 end
 
 # Inverse pass 
 function inverse(X::AbstractArray{T, N}, Y::AbstractArray{T, N}, S::SummarizedNet) where {T, N}
-    S.cond_net.inverse(X,  Y)
+    X = S.cond_net.inverse(X,  Y)
+    S.squeezer.inverse(X)
 end
 
 # Backward pass and compute gradients
 function backward(ΔX::AbstractArray{T, N}, X::AbstractArray{T, N}, Y::AbstractArray{T, N}, S::SummarizedNet; Y_save=nothing) where {T, N}
 	ΔX, X, ΔY = S.cond_net.backward(ΔX,X,Y)
+    Y_save = S.squeezer.forward(Y_save)
     ΔY = S.sum_net.backward(ΔY, Y_save)
     return ΔX, X, ΔY
 end
