### 2024-05-28 #6554 @rafaelorozco
Hello @Nando-Hegemann thank you for your useful comments and feedback! I will be integrating them over the course of the next few days.

For anyone reading this thread, I just want to clarify that we never claim to this software performs faster than non-invertiblity-exploiting frameworks. In fact, due to recomputing the intermediate activations this package should use more compute time. The main bottleneck in normalizing flows has been memory efficiency (cited in the paper) so this is a fair price to pay for enabling inference on large inputs and deep architectures.
