+++ docs/paper/paper.bib
@@ -1,3 +1,10 @@
+@article{alemohammad2023self,
+  title={Self-consuming generative models go mad},
+  author={Alemohammad, Sina and Casco-Rodriguez, Josue and Luzi, Lorenzo and Humayun, Ahmed Imtiaz and Babaei, Hossein and LeJeune, Daniel and Siahkoohi, Ali and Baraniuk, Richard G},
+  journal={arXiv preprint arXiv:2307.01850},
+  year={2023}
+}
+
 @article{peters2022point,
   title={Point-TO-SET DISTANCE FUNCTIONS FOR OUTPUT-CONSTRAINED NEURAL NETWORKS.},
   author={Peters, Bas},
