### 2024-05-28 9323ad3
add memory benchmarks
