+++ examples/applications/application_conditional_hint_seismic.jl
@@ -1,162 +1,0 @@
-# Generative model using the change of variables formula
-# Author: Philipp Witte, pwitte3@gatech.edu
-# Date: January 2020
-
-using LinearAlgebra, InvertibleNetworks, PyPlot, Flux, Random, Test, JLD, Statistics
-import Flux.Optimise.update!
-
-# Random seed
-Random.seed!(66)
-
-
-####################################################################################################
-
-# Load original data X (size of n1 x n2 x nc x ntrain)
-datadir = dirname(pathof(InvertibleNetworks))*"/../data/"
-filename = "seismic_samples_64_by_64_num_10k.jld"
-~isfile("$(datadir)$(filename)") && run(`curl -L https://www.dropbox.com/s/mh5dv0yprestot4/seismic_samples_64_by_64_num_10k.jld\?dl\=0 --create-dirs -o $(datadir)$(filename)`)
-X_orig = load("$(datadir)$(filename)")["X"]
-n1, n2, nc, nsamples = size(X_orig)
-AN = ActNorm(nsamples)
-X_orig = AN.forward(X_orig) # zero mean and unit std
-
-# Split in training - testing
-ntrain = Int(nsamples*.9)
-ntest = nsamples - ntrain
-
-# Dimensions after wavelet squeeze to increase no. of channels
-nx = Int(n1/2)
-ny = Int(n2/2)
-n_in = Int(nc*4)
-
-# Apply wavelet squeeze (change dimensions to -> n1/2 x n2/2 x nc*4 x ntrain)
-X_train = zeros(Float32, nx, ny, n_in, ntrain)
-for j=1:ntrain
-    X_train[:, :, :, j:j] = wavelet_squeeze(X_orig[:, :, :, j:j])
-end
-
-X_test = zeros(Float32, nx, ny, n_in, ntest)
-for j=1:ntest
-    X_test[:, :, :, j:j] = wavelet_squeeze(X_orig[:, :, :, ntrain+j:ntrain+j])
-end
-
-# Create network
-n_hidden = 64
-batchsize = 4
-depth = 8
-CH = NetworkConditionalHINT(n_in, n_hidden, depth)
-Params = get_params(CH)
-
-####################################################################################################
-
-# Loss
-function loss(CH, X, Y)
-    Zx, Zy, logdet = CH.forward(X, Y)
-    f = -log_likelihood(tensor_cat(Zx, Zy)) - logdet
-    ΔZ = -∇log_likelihood(tensor_cat(Zx, Zy))
-    ΔZx, ΔZy = tensor_split(ΔZ)
-    ΔX, ΔY = CH.backward(ΔZx, ΔZy, Zx, Zy)[1:2]
-    return f, ΔX, ΔY
-end
-
-# Training
-maxiter = 1000
-opt = Flux.ADAM(1f-3)
-lr_step = 100
-lr_decay_fn = Flux.ExpDecay(1f-3, .9, lr_step, 0.)
-fval = zeros(Float32, maxiter)
-
-for j=1:maxiter
-
-    # Evaluate objective and gradients
-    idx = randperm(ntrain)[1:batchsize]
-    X = X_train[:, :, :, idx]
-    Y = X + .5f0*randn(Float32, nx, ny, n_in, batchsize)
-    
-    fval[j] = loss(CH, X, Y)[1]
-    mod(j, 10) == 0 && (print("Iteration: ", j, "; f = ", fval[j], "\n"))
-
[…]
