+++ examples/applications/conditional_sampling/amortized_glow_mnist_inpainting.jl
@@ -1,16 +1,20 @@
-using Pkg
-Pkg.activate(".")
-
 # Take around 6 minutes on CPU
+using Pkg
+
+Pkg.add("InvertibleNetworks")
+Pkg.add("Flux")
+Pkg.add("MLDatasets")
+Pkg.add("ProgressMeter")
+Pkg.add("MLUtils")
+Pkg.add("ImageTransformations")
+
 using InvertibleNetworks
 using Flux
 using LinearAlgebra
 using MLDatasets
-using Statistics
-using PyPlot
-using ProgressMeter: Progress, next!
-using Images
+using ImageTransformations
 using MLUtils
+using ProgressMeter: Progress, next!
 
 function posterior_sampler(G, y, size_x; device=gpu, num_samples=1, batch_size=16)
     # make samples from posterior for train sample 
@@ -111,52 +115,35 @@ for e=1:epochs # epoch loop
     append!(loss_val, norm(ZX)^2 / (N*n_val) - logdet_i / N)  # normalize by image size and batch size
 end
 
-# Training logs
-final_obj_train = round(loss_train[end];digits=3)
-final_obj_val = round(loss_val[end];digits=3)
 
-fig = figure()
-title("Objective value: train=$(final_obj_train) validation=$(final_obj_val)")
+Pkg.add("Statistics")
+Pkg.add("Plots")
+using Statistics
+using Plots
+
+# Training logs
 plot(loss_train;label="Train"); 
 plot(batches:batches:batches*(epochs), loss_val;label="Validation"); 
-xlabel("Parameter update"); ylabel("Negative log likelihood objective") ;
-legend()
-savefig("log.png",dpi=300)
+xlabel!("Parameter update"); ylabel!("Negative log likelihood objective") ;
+savefig("log.png")
 
-# Make Figure of README
-num_plot = 2
-fig = figure(figsize=(11, 5));
-for (i,ind) in enumerate([1,3])
-	x = XY_val[1][:,:,:,ind:ind]
-	y = XY_val[2][:,:,:,ind:ind]
-	X_post = posterior_sampler(G, y, size(x); device=device, num_samples=64) |> cpu
+# Make Figure like README
+ind = 1
+x = XY_val[1][:,:,:,ind:ind]
+y = XY_val[2][:,:,:,ind:ind]
+X_post = posterior_sampler(G, y, size(x); device=device, num_samples=64) |> cpu
 
-	X_post_mean = mean(X_post; dims=ndims(X_post)) 
-	X_post_var  = var(X_post;  dims=ndims(X_post)) 
+X_post_mean = mean(X_post; dims=ndims(X_post)) 
+X_post_var  = var(X_post;  dims=ndims(X_post)) 
 
-	ssim_val = round(assess_ssim(X_post_mean[:,:,1,1], x[:,:,1,1]) ,digits=2)
+p1 = heatmap(x[:,:,1,1];title="Ground truth",cbar=false,clims = (0, 1),axis=([], false))
+p2 = heatmap(y[:,:,1,1];title="Observation",cbar=false,clims = (0, 1),axis=([], false))
+p3 = heatmap(X_post_mean[:,:,1,1];title="Posterior mean",cbar=false,clims = (0, 1),axis=([], false))
+p4 = heatmap(X_post_var[:,:,1,1];title="Posterior variance",cbar=false,axis=([], false))
[…]
