### 2024-05-30 a4d7f2d
Create CONTRIBUTING.md
