### 2024-05-30 cf19a4f
Update README.md w/ contribution guide
