### 2024-05-30 slimgroup/InvertibleNetworks.jl#106 @rafaelorozco
[Update README.md w/ contribution guide](https://github.com/slimgroup/InvertibleNetworks.jl/commit/cf19a4f3411fa1b5fd83c08b7fd465f81a6e9a95)

Added a contribution guide to the readme. Hope you agree this is good!
