### 2024-05-30 slimgroup/InvertibleNetworks.jl#106 @aurorarossi
Yes, I close the issue.
