### 2024-06-03 a5ee3be
Removed deps restrictions
