+++ Project.toml
@@ -16,12 +16,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Wavelets = "29a6e085-ba6d-5f35-a997-948ac2efa89a"
 
 [compat]
-CUDA = "1, 2, 3, 4, 5"
-cuDNN = "1"
-ChainRulesCore = "0.8, 0.9, 0.10, 1"
+# CUDA = "1, 2, 3, 4, 5"
+# cuDNN = "1"
+# ChainRulesCore = "0.8, 0.9, 0.10, 1"
 Flux = "0.11, 0.12, 0.13, 0.14"
 JOLI = "0.7, 0.8"
-NNlib = "0.7, 0.8, 0.9"
+# NNlib = "0.7, 0.8, 0.9"
 TimerOutputs = "0.5"
 Wavelets = "0.9, 0.10"
 julia = "1"
