+++ Project.toml
@@ -6,7 +6,6 @@ version = "2.2.8"
 [deps]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
-cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 JOLI = "bb331ad6-a1cf-11e9-23da-9bcb53c69f6f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -14,17 +13,11 @@ NNlib = "872c559c-99b0-510c-b3b7-b6c96a88d5cd"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Wavelets = "29a6e085-ba6d-5f35-a997-948ac2efa89a"
+cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
 
 [compat]
-# CUDA = "1, 2, 3, 4, 5"
-# cuDNN = "1"
-# ChainRulesCore = "0.8, 0.9, 0.10, 1"
-Flux = "0.11, 0.12, 0.13, 0.14"
-JOLI = "0.7, 0.8"
-# NNlib = "0.7, 0.8, 0.9"
 TimerOutputs = "0.5"
 Wavelets = "0.9, 0.10"
-julia = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
