+++ src/conditional_layers/conditional_layer_glow.jl
@@ -60,7 +60,8 @@ or
 """
 struct ConditionalLayerGlow <: NeuralNetLayer
     C::Conv1x1
-    RB::ResidualBlock
+    #RB::ResidualBlock
+    RB::FluxBlock
     logdet::Bool
     activation::ActivationFunction
 end
@@ -74,11 +75,17 @@ function ConditionalLayerGlow(C::Conv1x1, RB::ResidualBlock; logdet=false, activ
 end
 
 # Constructor from input dimensions
-function ConditionalLayerGlow(n_in::Int64, n_cond::Int64, n_hidden::Int64;freeze_conv=false, k1=3, k2=1, p1=1, p2=0, s1=1, s2=1, logdet=false, activation::ActivationFunction=SigmoidLayer(), rb_activation::ActivationFunction=RELUlayer(), ndims=2)
+function ConditionalLayerGlow(n_in::Int64, n_cond::Int64, n_hidden::Int64;Unet_t=nothing,freeze_conv=false, k1=3, k2=1, p1=1, p2=0, s1=1, s2=1, logdet=false, activation::ActivationFunction=SigmoidLayer(), rb_activation::ActivationFunction=RELUlayer(), ndims=2)
 
     # 1x1 Convolution and residual block for invertible layers
     C  = Conv1x1(n_in; freeze=freeze_conv)
-    RB = ResidualBlock(Int(n_in/2)+n_cond, n_hidden; n_out=n_in, activation=rb_activation, k1=k1, k2=k2, p1=p1, p2=p2, s1=s1, s2=s2, fan=true, ndims=ndims)
+    #RB = ResidualBlock(Int(n_in/2)+n_cond, n_hidden; unet=true, n_out=n_in, activation=rb_activation, k1=k1, k2=k2, p1=p1, p2=p2, s1=s1, s2=s2, fan=true, ndims=ndims)
+    
+    unet_lev = 2
+    h = Chain(Unet_t(Int(n_in/2)+n_cond, n_in, unet_lev;dims=2));
+    trainmode!(h, true);  #can we get rid of this? NO!!!!
+    RB = FluxBlock(h) #|> device;
+
 
     return ConditionalLayerGlow(C, RB, logdet, activation)
 end
