### 2024-06-17 #6554 @rafaelorozco
Hello things are going well, I have been chipping away at the recommendations made by the reviewers.

I added the contribution guidelines as requested.

I added some scripts to reproduce the benchmarks in the manuscript https://github.com/slimgroup/InvertibleNetworks.jl/pull/109

I also changed some of the dependencies so that it is easier to install the package. @Nando-Hegemann was not able to install the package although I dont know what the error was so I took away pyplots as a dependency from the first example which might cause problems. If you are still getting errors please let me know what they are.
