### 2024-06-24 313192f
can take away immediate donwsampling
