### 2024-06-25 3dd0c66
add statistics again
