### 2024-06-25 2eb08ad
add immediate downsamp
