### 2024-07-05 dabe5b6
add context to paper, fix dois
