+++ docs/paper/paper.bib
@@ -1,3 +1,30 @@
+@inproceedings{louboutin2022accelerating,
+  title={Accelerating innovation with software abstractions for scalable computational geophysics},
+  author={Louboutin, Mathias and Witte, Philipp and Siahkoohi, Ali and Rizzuti, Gabrio and Yin, Ziyi and Orozco, Rafael and Herrmann, Felix J},
+  booktitle={Second International Meeting for Applied Geoscience \& Energy},
+  pages={1482--1486},
+  year={2022},
+  organization={Society of Exploration Geophysicists and American Association of Petroleum~…}
+}
+
+@misc{NormalizingFlows.jl,
+	author  = {Zuheng Xu, Xianda Sun, Tor Erlend Fjelde, Hong Ge and contributors},
+	title   = {NormalizingFlows.jl},
+	url     = {https://github.com/TuringLang/NormalizingFlows.jl},
+	version = {v0.1.0},
+	year    = {2023},
+	month   = {6}
+}
+
+@inproceedings{fjelde2020bijectors,
+  title={Bijectors. jl: Flexible transformations for probability distributions},
+  author={Fjelde, Tor Erlend and Xu, Kai and Tarek, Mohamed and Yalburgi, Sharan and Ge, Hong},
+  booktitle={Symposium on Advances in Approximate Bayesian Inference},
+  pages={1--17},
+  year={2020},
+  organization={PMLR}
+}
+
 @article{alemohammad2023self,
   title={Self-consuming generative models go mad},
   author={Alemohammad, Sina and Casco-Rodriguez, Josue and Luzi, Lorenzo and Humayun, Ahmed Imtiaz and Babaei, Hossein and LeJeune, Daniel and Siahkoohi, Ali and Baraniuk, Richard G},
@@ -11,6 +38,7 @@
   journal={Journal of Applied \& Numerical Optimization},
   volume={4},
   number={2},
+  doi  = {10.23952/jano.4.2022.2.05},
   year={2022}
 }
 
@@ -48,6 +76,7 @@
   number={4},
   pages={1452--1466},
   year={2020},
+  doi={10.1109/tnnls.2020.3042395},
   publisher={IEEE}
 }
 
@@ -127,6 +156,7 @@
   title={Zur theorie der orthogonalen funktionensysteme},
   author={Haar, Alfred},
   year={1909},
+  doi={10.1007/bf01456927},
   publisher={Georg-August-Universitat, Gottingen.}
 }
 
@@ -162,6 +192,7 @@
   volume={35},
   number={9},
   pages={8191--8199},
+  doi={10.1609/aaai.v35i9.16997},
   year={2021}
 }
 
@@ -177,6 +208,7 @@
   title={normflows: A PyTorch Package for Normalizing Flows},
   author={Stimper, Vincent and Liu, David and Campbell, Andrew and Berenz, Vincent and Ryll, Lukas and Sch{\"o}lkopf, Bernhard and Hern{\'a}ndez-Lobato, Jos{\'e} Miguel},
   journal={arXiv preprint arXiv:2302.12014},
+  doi={10.21105/joss.05361},
   year={2023}
 }
 
@@ -195,6 +227,7 @@
   booktitle={SEG International Exposition and Annual Meeting},
   pages={D031S057R006},
   year={2020},
+  doi={10.1190/segam2020-3428150.1},
   organization={SEG}
 }
 
@@ -226,6 +259,7 @@
   volume={42},
   number={7},
   pages={474--486},
+  doi={10.1190/tle42070474.1},
   year={2023},
   publisher={Society of Exploration Geophysicists}
 }
@@ -244,6 +278,7 @@
   volume={88},
   number={3},
[…]
