+++ docs/paper/paper.bib
@@ -4,6 +4,7 @@
   booktitle={Second International Meeting for Applied Geoscience \& Energy},
   pages={1482--1486},
   year={2022},
+  doi={10.1190/image2022-3750561.1},
   organization={Society of Exploration Geophysicists and American Association of Petroleum~…}
 }
 
@@ -29,6 +30,7 @@
   title={Self-consuming generative models go mad},
   author={Alemohammad, Sina and Casco-Rodriguez, Josue and Luzi, Lorenzo and Humayun, Ahmed Imtiaz and Babaei, Hossein and LeJeune, Daniel and Siahkoohi, Ali and Baraniuk, Richard G},
   journal={arXiv preprint arXiv:2307.01850},
+  doi={10.52591/lxai202312101},
   year={2023}
 }
 
@@ -50,6 +52,7 @@
   number={1},
   pages={65--98},
   year={2017},
+  doi={10.1137/141000671},
   publisher={SIAM},
   url={https://doi.org/10.1137/141000671}
 }
@@ -58,6 +61,7 @@
   title={Symmetric block-low-rank layers for fully reversible multilevel neural networks},
   author={Peters, Bas and Haber, Eldad and Lensink, Keegan},
   journal={arXiv preprint arXiv:1912.12137},
+  doi={10.48550/arXiv.1912.12137},
   year={2019}
 }
 
@@ -65,6 +69,7 @@
   title={Memory Efficient Invertible Neural Networks for 3D Photoacoustic Imaging},
   author={Orozco, Rafael and Louboutin, Mathias and Herrmann, Felix J},
   journal={arXiv preprint arXiv:2204.11850},
+  doi={10.48550/arXiv.2204.11850},
   year={2022}
 }
 
@@ -84,6 +89,7 @@
   title={A differentiable programming system to bridge machine learning and scientific computing},
   author={Innes, Mike and Edelman, Alan and Fischer, Keno and Rackauckas, Chris and Saba, Elliot and Shah, Viral B and Tebbutt, Will},
   journal={arXiv preprint arXiv:1907.07587},
+  doi={10.48550/arXiv.1907.07587},
   year={2019}
 }
 
@@ -146,10 +152,15 @@
   url          = {https://doi.org/10.5281/zenodo.4296287}
 }
 
-@article{paszke2017automatic,
-  title={Automatic differentiation in pytorch},
-  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  year={2017}
+
+
+@article{paszke2019pytorch,
+  title={Pytorch: An imperative style, high-performance deep learning library},
+  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  journal={Advances in neural information processing systems},
+  volume={32},
+  doi={doi.org/10.48550/arXiv.1912.01703},
+  year={2019}
 }
 
 @book{haar1909theorie,
@@ -164,6 +175,7 @@
   title={Nice: Non-linear independent components estimation},
   author={Dinh, Laurent and Krueger, David and Bengio, Yoshua},
   journal={arXiv preprint arXiv:1410.8516},
+  doi={10.48550/arXiv.1410.8516},
   year={2014}
 }
 
@@ -171,6 +183,7 @@
   title={Density estimation using real nvp},
[…]
