+++ docs/paper/paper.bib
@@ -159,7 +159,7 @@
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   journal={Advances in neural information processing systems},
   volume={32},
-  doi={doi.org/10.48550/arXiv.1912.01703},
+  doi={10.48550/arXiv.1912.01703},
   year={2019}
 }
 
