### 2024-07-05 #6554 @rafaelorozco
Hello all, here are my responses and integrations of the comments from @Nando-Hegemann

> **General Checks -> Reproducibility**: I'm not able to reproduce the results of the paper (or run any of the other examples). I'm aware that this is due to my lack of familiarity with the Julia language, but maybe you could include a simple section on "How to run the MNIST example" (simple command line instructions) in the README/Doc as well as a description on how to set up an environment to execute the examples/tests. Right now, after freshly installing Julia and running the scripts/tests I simply get LoadError(s).

It was not clear what the error you are refering to was so I tried running the example from various systems and arrived to the conclusion that the pyplots dependency might be an issue if the system python does not have matplotlib installed. To resolve this I took away that dependency from the example and instead use Plots.jl. I also add lines to manually add the package dependencies in case new Julia users do not know how to add those.
>
> **Functionality -> Functionality** & **Functionality -> Performance**: I agree with @aurorarossi that a script reproducing the plots of the paper would be good. A simple (python) script and respective environment for the python setup used to produce the results would be a good addition.

Great idea! I added the benchmark scripts in this PR https://github.com/slimgroup/InvertibleNetworks.jl/pull/109

> **Documentation -> Example usage**:
>
> 1. You have a lot of example scripts of which only a few are included in the online documentation (which is fine). However, the examples in the online documentation on provide the source code of the scripts and an image. I think providing more context to the examples (what is the setting, why is it interesting as an example) and describing how the individual steps work and what the output picture depicts would be good.

Thank you for the feedback, I completely agree and I added some more comments to the examples in the online documentation

> 2. Currently you copy-pasted source code from your scripts to the examples in the doc. To decrease maintenance issues, I'd suggest referencing (certain lines of) the Julia scripts in the doc, so that the doc always shows source code from executable files. This way your doc and the example scripts don't run out of sync if you change something.
>
> With .rst files you could do something like
>
> ```
> .. literalinclude:: ../examples/applications/file_name.jl
>   :language: julia
>   :lines: 8-9
> ```

This is a very useful suggestion. Currently those example are being run as code during documentation generation and I was unable to get that working in tandem with the .rst files. But I will keep an eye out in the future to see if we can integrate both of these methods.

> **Documentation -> Community Guidelines**: Again, I agree with @aurorarossi to include a clearly visible link to your general contribution guidelines in your README.

Thank you for the feedback we have added some guidelines.

> **Software Paper -> State of the field**: In the paper you describe that your software performs faster then other packages such as FrEIA or normflows, but I'd suggest adding some additional notes on what separates this software from other packages. For example you could focus on the following:
>
> 1. Is there a reason to implement Invertible Networks in Julia rather than in python?

From the onset, our goal was interoperability of this package with other packages in the workflow of our lab. The multiple dispatch system of Julia as a fundamental design feature greatly aided this goal for reference please see this publication https://arxiv.org/pdf/2203.15038. Other than this programatic reason, we are interested in scalable software since our goal is to offer solutions for imaging problems that have high degrees of freedom and require efficient solutions. I have added some sentences to describe these reasons.

> 2. Does your software include algorithms/architectures/training routines not implemented anywhere else?
> 3. Does your software combine established algorithms from multiple other packages, thus providing a more versatile and easy to use framework?

Here we are only claiming to uniquely have access to memory efficiency that has enabled its application in large scale imaging problems. This has been a huge bottleneck for the acceptance of normalizing flows so we believe it is important for the community.

>    In other words, you should reason why you implemented a new software package. If you mainly compute gradients analytically instead of using autograd one could argue that it would be better to integrate these computational improvements in other already existing packages.

Im afraid it is a bit difficult to phrase these arguments since this package was implemented 4 years ago when there were few options for existing INN packages. It is now has mature functionality that has been used for many publications.

> Also, are there any other Julia packages for INNs? If so they should be mentioned and differences should be explained as well.

Some packages that come to mind are: Bijectors.jl, NormalizingFlows.jl. The important difference being that they do not implement the manually defined gradients required for memory efficiency of large scale training.  I have added them to the writeup.

> **Software Paper -> Quality of writing**: I have only some minor formulation changes and suggestions. See [slimgroup/InvertibleNetworks.jl#108](https://github.com/slimgroup/InvertibleNetworks.jl/pull/108).

Thank you for the suggestions, I have incorporated them into the writeup.
