### 2024-07-24 caeb7f0
Make minor readability suggestions
