### 2024-07-24 #6554 @drvinceknight
> I'm a bit confused, did @Nando-Hegemann check the last changes? if I'm not mistaken I see his last activity in this PR on May 22nd.

Thanks @aurorarossi, given the nature of his comments I am satisfied that they have been addressed. @rafaelorozco would you agree?
