### 2024-07-24 #6554 @drvinceknight
@rafaelorozco I have made some minor suggestions here: https://github.com/slimgroup/InvertibleNetworks.jl/pull/111
