### 2024-07-24 363ca49
Update Project.toml
bump version for joss paper
