### 2024-07-29 #6554 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
