### 2024-07-29 17b3bdc
fix cond layer init and backward
