### 2024-07-30 5035ce8
minor changes in paper.md
