### 2024-07-30 303fdd4
minor changes in paper.bib
