+++ docs/paper/paper.bib
@@ -20,7 +20,7 @@
 @inproceedings{fjelde2020bijectors,
   title={Bijectors. jl: Flexible transformations for probability distributions},
   author={Fjelde, Tor Erlend and Xu, Kai and Tarek, Mohamed and Yalburgi, Sharan and Ge, Hong},
-  booktitle={Symposium on Advances in Approximate Bayesian Inference},
+  booktitle={Symposium on Advances in Approximate {B}ayesian Inference},
   pages={1--17},
   year={2020},
   organization={PMLR}
@@ -133,7 +133,7 @@
 
 @software{freia,
   author = {Ardizzone, Lynton and Bungert, Till and Draxler, Felix and Köthe, Ullrich and Kruse, Jakob and Schmier, Robert and Sorrenson, Peter},
-  title = {{Framework for Easily Invertible Architectures (FrEIA)}},
+  title = {Framework for Easily Invertible Architectures ({FrEIA})},
   year = {2018-2022},
   url = {https://github.com/vislearn/FrEIA}
 }
@@ -155,7 +155,7 @@
 
 
 @article{paszke2019pytorch,
-  title={Pytorch: An imperative style, high-performance deep learning library},
+  title={Py{T}orch: An imperative style, high-performance deep learning library},
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   journal={Advances in neural information processing systems},
   volume={32},
@@ -200,7 +200,7 @@
 }
 
 @inproceedings{kruse2021hint,
-  title={HINT: Hierarchical invertible neural transport for density estimation and Bayesian inference},
+  title={HINT: Hierarchical invertible neural transport for density estimation and {B}ayesian inference},
   author={Kruse, Jakob and Detommaso, Gianluca and K{\"o}the, Ullrich and Scheichl, Robert},
   booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
   volume={35},
@@ -255,7 +255,7 @@
 }
 
 @article{siahkoohi2022wave,
-  title={Wave-equation-based inversion with amortized variational Bayesian inference},
+  title={Wave-equation-based inversion with amortized variational {B}ayesian inference},
   author={Siahkoohi, Ali and Orozco, Rafael and Rizzuti, Gabrio and Herrmann, Felix J},
   journal={arXiv preprint arXiv:2203.15881},
   doi={10.48550/arXiv.2203.15881},
@@ -330,7 +330,7 @@
 }
 
 @inproceedings{orozco2023adjoint,
-  title={Adjoint operators enable fast and amortized machine learning based Bayesian uncertainty quantification},
+  title={Adjoint operators enable fast and amortized machine learning based {B}ayesian uncertainty quantification},
   author={Orozco, Rafael and Siahkoohi, Ali and Rizzuti, Gabrio and van Leeuwen, Tristan and Herrmann, Felix J},
   booktitle={Medical Imaging 2023: Image Processing},
   volume={12464},
@@ -341,7 +341,7 @@
 }
 
 @article{khorashadizadeh2023conditional,
-  title={Conditional injective flows for Bayesian imaging},
+  title={Conditional injective flows for {B}ayesian imaging},
   author={Khorashadizadeh, AmirEhsan and Kothari, Konik and Salsi, Leonardo and Harandi, Ali Aghababaei and de Hoop, Maarten and Dokmani{\'c}, Ivan},
   journal={IEEE Transactions on Computational Imaging},
   volume={9},
