### 2024-07-30 #6554 @danielskatz
@rafaelorozco - I'm suggesting the changes in https://github.com/slimgroup/InvertibleNetworks.jl/pull/113 - please merge this, or let me know what you disagree with, then we can continue the acceptance process
