### 2024-07-30 #6554 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
