### 2024-07-30 #6554 @danielskatz
@rafaelorozco - can you confirm that you don't want to have an acknowledgements section, and that there's no funding or anything else you need to acknowledge?
