### 2024-07-30 #6554 @danielskatz
Congratulations to @rafaelorozco (Rafael Orozco) and co-authors on your publication!!

JOSS depends on volunteers and couldn't succeed without you
